Fix Author folder structure
