+++ apps/codeforafrica/src/components/Author/Author.snap.js
@@ -1,0 +1,19 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<Author /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-3vtd9l"
+  >
+    <span
+      class="MuiTypography-root MuiTypography-body2 css-16etphd-MuiTypography-root"
+    >
+      Article by
+       
+    </span>
+    <div
+      class="MuiBox-root css-hjoi4d"
+    />
+  </div>
+</div>
+`;
