+++ apps/codeforafrica/src/components/Author/Author.test.js
@@ -1,7 +1,7 @@
 import { createRender } from "@commons-ui/testing-library";
 import React from "react";
 
-import Author from ".";
+import Author from "./Author";
 
 import theme from "@/codeforafrica/theme";
