--- apps/codeforafrica/src/components/Author/index.snap.js
@@ -1,51 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<ArticleAuthorCopy /> renders unchanged 1`] = `
-<div>
-  <div
-    class="MuiBox-root css-d944bl"
-  >
-    <span
-      class="MuiTypography-root MuiTypography-body2 css-lupkoz-MuiTypography-root"
-    >
-      Article by
-       
-    </span>
-    <div
-      class="MuiBox-root css-139x83s"
-    >
-      <span
-        class="MuiTypography-root MuiTypography-body1SemiBold css-1s0xknt-MuiTypography-root"
-      />
-      <span
-        class="MuiTypography-root MuiTypography-body1 css-4tj74i-MuiTypography-root"
-      />
-    </div>
-  </div>
-</div>
-`;
-
-exports[`<Author /> renders unchanged 1`] = `
-<div>
-  <div
-    class="MuiBox-root css-d944bl"
-  >
-    <span
-      class="MuiTypography-root MuiTypography-body2 css-lupkoz-MuiTypography-root"
-    >
-      Article by
-       
-    </span>
-    <div
-      class="MuiBox-root css-139x83s"
-    >
-      <span
-        class="MuiTypography-root MuiTypography-body1SemiBold css-1s0xknt-MuiTypography-root"
-      />
-      <span
-        class="MuiTypography-root MuiTypography-body1 css-4tj74i-MuiTypography-root"
-      />
-    </div>
-  </div>
-</div>
-`;
